Move token stuff into its own package.
