+++ go/conversation/conversation_views.py
@@ -13,7 +13,7 @@ from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 
 from go.vumitools.exceptions import ConversationSendError
-from go.base.django_token_manager import DjangoTokenManager
+from go.token.django_token_manager import DjangoTokenManager
 from go.conversation.forms import (
     ConversationForm, ConversationGroupForm, ConfirmConversationForm,
     ReplyToMessageForm)
