+++ go/conversation/views.py
@@ -1,22 +1,15 @@
-import urlparse
 from urllib import urlencode
 
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import render, redirect, Http404
+from django.shortcuts import render, redirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.views.decorators.http import require_POST
 from django.contrib import messages
 from django.http import HttpResponse
-from django.conf import settings
-from django.contrib.auth.views import logout
-from django.core.urlresolvers import reverse
-
-from vumi.persist.redis_manager import RedisManager
 
 from go.conversation.forms import ConversationSearchForm
 from go.base.utils import get_conversation_view_definition, conversation_or_404
 from go.conversation_tmp.forms import CampaignGeneralForm
-from go.vumitools.token_manager import TokenManager
 
 
 CONVERSATIONS_PER_PAGE = 12
@@ -129,37 +122,3 @@ def new_conversation(request):
 
     return redirect(view_def.get_view_url(
         next_view, conversation_key=conv.key))
-
-
-def token(request, token):
-    # This is special, since we don't necessarily have an authenticated user
-    # and we definitely don't have a conversation.
-    redis = RedisManager.from_config(settings.VUMI_API_CONFIG['redis_manager'])
-    tm = TokenManager(redis.sub_manager('token_manager'))
-    token_data = tm.get(token)
-    if not token_data:
-        raise Http404
-
-    user_id = int(token_data['user_id'])
-    redirect_to = token_data['redirect_to']
-    system_token = token_data['system_token']
-
-    # If we're authorized and we're the same user_id then redirect to
-    # where we need to be
-    if not user_id or request.user.id == user_id:
-        path, _, qs = redirect_to.partition('?')
-        params = urlparse.parse_qs(qs)
-        # since the token can be custom we prepend the size of the user_token
-        # to the token being forwarded so the view handling the `redirect_to`
-        # can lookup the token and verify the system token.
-        params.update({'token': '%s-%s%s' % (len(token), token, system_token)})
-        return redirect('%s?%s' % (path, urlencode(params)))
-
-    # If we got here then we need authentication and the user's either not
-    # logged in or is logged in with a wrong account.
-    if request.user.is_authenticated():
-        logout(request)
-        messages.info(request, 'Wrong account for this token.')
-    return redirect('%s?%s' % (reverse('auth_login'), urlencode({
-        'next': reverse('token', kwargs={'token': token}),
-        })))
