+++ go/token/urls.py
@@ -1,0 +1,9 @@
+from django.conf.urls.defaults import patterns, url
+
+from go.token import views
+
+
+urlpatterns = patterns('',
+    url(r'^task/$', views.token_task, name='token_task'),
+    url(r'^(?P<token>\w+)/$', views.token, name='token'),
+)
